[frontend] improvements
